+++ react-rs/src/components/CharacterList.tsx
@@ -1,3 +1,4 @@
+"use client"; 
 import React from "react";
 import Link from "next/link";
 import { useAppDispatch, useAppSelector } from "../store/store";
