+++ react-rs/src/context/ThemeContext.tsx
@@ -1,3 +1,4 @@
+"use client"; 
 import React, { createContext, useState, ReactNode } from "react";
 
 interface ThemeContextType {
