+++ react-rs/src/hooks/useTheme.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useContext } from "react";
 import { ThemeContext } from "../context/ThemeContext";
