+++ react-rs/src/store/store.ts
@@ -1,3 +1,4 @@
+"use client"; 
 import { configureStore, createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
